rework blpConnect
