+++ R/blpapi.R
@@ -1,15 +1,3 @@
-blpConnect <- function(host=getOption("blpHost", "localhost"),
-                       port=getOption("blpPort", 8194L),
-                       logfile) {
-    if (missing(logfile)) {
-        logfile <- file.path("/tmp/",
-                             paste0("blpapi_", format(Sys.time(), "%Y%m%d_%H%M%S"),
-                                    "_", Sys.getpid(), ".log"))
-    }
-    if (storage.mode(port) != "integer") port <- as.integer(port)
-    if (storage.mode(host) != "character") stop("Host argument must be character.", call.=FALSE)
-    .Call("bdp_connect", host, port, logfile, PACKAGE="Rblpapi")
-}
 
 blpDisconnect <- function(conn) {
     # do nothing, just return a simple test
